Adding ability to delete rows.
In the course of this, the way that grid rows are collected have changed. Now we just run numerically from 1-100 instead of incrementing up and bailing if there is a break in the numerical chain. Obviously this was flawed since deleting a row will break the chain causing all of the rows after it to dissapear.